Contact List with Firebase, Styling and Output
Implemented the Contact List to view all contacts. When a firebase
connection is established, the component fetches all information from
the 'contacts' collection and shows a list. Styling with SCSS also done
with width 100% to be implemented in a flex/grid.
When clicked on a item, the component emits a string output, which is
the Id of the collection entry.